+++ CHANGES.md
@@ -8,11 +8,18 @@
   - added AddCustomItem
   - added AddCustomItemsFromJson
 - added ItemCreator
+## Recipe
+- added customRecipeDatabase
+  - added AddCustomRecipe
+  - added AddCustomRecipesFromJson
+- added RecipeCreator
 ## Patches
 - added ItemDatabase.DeSerialize Postfix
 - added ItemOperations.SetCollectibleItemValues Postfix
 - added Item.ItemAppearance.LoadSprite Prefix
 - added Item.ItemAppearance.Loadprefab Prefix
+- added RecipeDatabase.DeSerialize Postfix
+- added Recipe.RecipeAppearance.LoadSprite Prefix
 
 # v0.1.0 (December 8th 2024)
 ## Furniture